Create rc.local
I modified my rc.local in my /etc account for my root account to check that I had an IP address, then begin running the script in the background.  This will start your script automatically on reboot and check that network services are running first so that it does not error out.